avoid compile error
git-svn-id: https://svn.code.sf.net/p/kshowmail/svn/trunk@182 9bd4a48d-2b12-44a0-a1bc-49c5f8e4eea2

+++ src/filtersetupitem.cpp
@@ -148,7 +148,7 @@ void FilterSetupItem::save( )
   //write criteria list and number of criterias
   if( !criteriaList.empty() )
   {
-    configFilter->writeEntry( CONFIG_ENTRY_FILTER_CRITERIA_NUMBER, criteriaList.size() );
+    configFilter->writeEntry( CONFIG_ENTRY_FILTER_CRITERIA_NUMBER, (int)criteriaList.size() );
 
     int ctr = 0; //number of the current criteria
     FilterCriteriaList_Type::iterator it;
